Update patch set 1
Patch Set 1:

The current value we use on the live systems is 4096. The default of 1024 gave errors when we had ssl issues. All connections switched to using only one box when one queue was restarted, and the default of 1024 meant some connections failed. This exacerbated the underlying issue and made the situation worse.

Happy to make it configurable but 4096 is alleviating that issue without adding another option to hiera. This seems to be a common rabbitmq change btw.

Patch-set: 1